robotframework-python3/requirements.txt
# Requirements needed when generating releases. See BUILD.rst for details.
invoke >= 0.20
rellu >= 0.6
docutils >= 0.14
robotframework >= 3.1.1
robotframework-seleniumlibrary >= 4.2
robotframework-pageobjectlibrary >= 1.1
pytest >= 3.2.2
